col_numとrow表記部分の入れ替え、共通処理のモジュール化、クラス名の変更
