+++ 08.ls_object/ls.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 require 'optparse'
-require_relative './file_list'
-require_relative './detail_file_list'
+require_relative './dir_content'
+require_relative './dir_info'
 
 options = ARGV.getopts('a', 'l', 'r')
 dir_path = ARGV.empty? ? '.' : ARGV[0]
